% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{eigen-package}
\alias{eigen-package}
\alias{BioCor}
\title{Eigen package}
\description{
The purpose of this package is to implement the identity provided at
https://arxiv.org/pdf/1908.03795v3.pdf for easier and
faster calculation of eigenvectors.
}
