% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvector.R
\name{eigenvector}
\alias{eigenvector}
\title{Eigenvector from the eigenvalues}
\usage{
eigenvector(x, A)
}
\arguments{
\item{x}{Numeric vector of the eigenvalues}
}
\value{
The eigenvector for each eigenvalue
}
\description{
Eigenvector from the eigenvalues
}
\examples{
C <- matrix(c(1,2,3,2,5,6,3,6,10), 3, 3)
D <- matrix(c(1, 1, -1, 1, 3, 1, -1, 1, 3), 3, 3)
eigenvector(c(14.9330343736593, 1, 0.0669656263407531), C)
}
